requirements.txt
Flask
flask-cors
bcrypt
PyJWT
python-dotenv